backend/pyproject.toml
[project]
name = "oneliner-backend"
version = "0.1.0"
description = "OneLiner - 每天读几句话"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "tortoise-orm[asyncpg]",
    "aerich",
    "pydantic-settings",
    "pyjwt",
    "httpx",
    "uvicorn",
]

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "."
